+++ src/extension_consts.ml
@@ -36,6 +36,7 @@ module Commands = struct
   let construct = ocaml_prefixed "construct"
   let merlin_jump = ocaml_prefixed "jump"
   let search_by_type = ocaml_prefixed "search-by-type"
+  let navigate_typed_holes = ocaml_prefixed "navigate-typed-holes"
 end
 
 module Command_errors = struct